Centralize csrf middleware and config in a module
