+++ server/src/pages/shot/server.js
@@ -1,6 +1,5 @@
 const express = require("express");
-const csrf = require("csurf");
-const csrfProtection = csrf({cookie: {httpOnly: true, secure: true, sameSite: 'lax', key: '__Host-csrf'}});
+const { csrfProtection } = require("../../middleware/csrf");
 const { Shot } = require("../../servershot");
 const { notFound } = require("../../pages/not-found/server");
 const reactrender = require("../../reactrender");
